Move Discord functions out separately
